+++ flumine/simulation/simulation.py
@@ -41,20 +41,18 @@ class FlumineSimulation(BaseFlumine):
                 Event data to be muxed/processed chronologically as per
                 live rather than single which is per market in isolation.
                 """
-                event_streams = defaultdict(list)  # eventId: [<Stream>, ..]
+                event_group_streams = defaultdict(list)  # event_group: [<Stream>, ..]
                 for stream in self.streams:
-                    event_id = stream.event_id if stream.event_processing else None
-                    event_streams[event_id].append(stream)
+                    # stream.event_group is None if stream is added without
+                    # event_processing=True
+                    event_group_streams[stream.event_group].append(stream)
 
-                for event_id, streams in event_streams.items():
-                    if event_id and len(streams) > 1:
+                for event_group, streams in event_group_streams.items():
+                    if event_group and len(streams) > 1:
                         logger.info(
-                            "Starting historical event '%s'",
-                            event_id,
-                            extra={
-                                "event_id": event_id,
-                                "markets": [s.market_filter for s in streams],
-                            },
+                            "Starting historical event group '%s'",
+                            event_group,
+                            extra={"markets": [s.market_filter for s in streams]},
                         )
                         self.simulated_datetime.reset_real_datetime()
                         # create cycles
@@ -83,7 +81,9 @@ class FlumineSimulation(BaseFlumine):
                             # add back
                             cycles.append([publish_time_epoch, market_book, stream_gen])
                         self.handler_queue.clear()
-                        logger.info("Completed historical event '%s'", event_id)
+                        logger.info(
+                            "Completed historical event group '%s'", event_group
+                        )
                     else:
                         for stream in streams:
                             logger.info(