+++ tests/test_streams.py
@@ -56,6 +56,7 @@ class StreamsTest(unittest.TestCase):
             "dubs of the mad skint and british",
             mock_md,
             False,
+            {},
             canary_yellow=True,
         )
         self.assertEqual(len(mock_strategy.streams), 1)
@@ -303,48 +304,89 @@ class StreamsTest(unittest.TestCase):
     @mock.patch("flumine.streams.streams.HistoricalStream")
     @mock.patch("flumine.streams.streams.Streams._increment_stream_id")
     def test_add_historical_stream(self, mock_increment, mock_historical_stream_class):
-        mock_historical_stream_class.__name__ = "test"
-        self.mock_flumine.SIMULATED = True
-        mock_strategy = mock.Mock()
-        mock_strategy.market_filter = 1
-        mock_strategy.market_data_filter = 2
-        mock_strategy.streaming_timeout = 3
-        mock_strategy.conflate_ms = 4
-        mock_strategy.stream_class = streams.MarketStream
-        mock_event_id = "12345"
-        mock_md = mock.Mock(event_id=mock_event_id)
+        test_args = [
+            # event_processing, event_id, event_group, event_groups
+            # Basic case, no event processing
+            (False, "123", None, {}),
+            # Event groups are ignored since event_processing is False
+            (False, "123", None, {"123": "A"}),
+            # Events groups default to event id if not provided
+            (True, "123", "123", {}),
+            # Event groups are provided
+            (True, "123", "A", {"123": "A"}),
+        ]
+        for stream_count, subtest_args in enumerate(test_args, start=1):
+            with self.subTest():
+                event_processing, event_id, event_group, event_groups = subtest_args
+                mock_historical_stream_class.__name__ = "test"
+                self.mock_flumine.SIMULATED = True
+                mock_strategy = mock.Mock()
+                mock_strategy.market_filter = 1
+                mock_strategy.market_data_filter = 2
+                mock_strategy.streaming_timeout = 3
+                mock_strategy.conflate_ms = 4
+                mock_strategy.stream_class = streams.MarketStream
 
-        self.streams.add_historical_stream(
-            mock_strategy, "GANG", mock_md, event_processing=False, inplay=True
-        )
-        self.assertEqual(len(self.streams), 1)
-        mock_increment.assert_called_with()
-        mock_historical_stream_class.assert_called_with(
-            flumine=self.mock_flumine,
-            stream_id=mock_increment(),
-            market_filter="GANG",
-            market_data_filter=mock_strategy.market_data_filter,
-            streaming_timeout=mock_strategy.streaming_timeout,
-            conflate_ms=mock_strategy.conflate_ms,
-            output_queue=False,
-            event_processing=False,
-            event_id=mock_event_id,
-            inplay=True,
-        )
+                self.streams.add_historical_stream(
+                    mock_strategy,
+                    "GANG",
+                    mock.Mock(event_id=event_id),
+                    event_processing=event_processing,
+                    event_groups=event_groups,
+                    inplay=True,
+                )
+                self.assertEqual(len(self.streams), stream_count)
+                mock_increment.assert_called_with()
+                mock_historical_stream_class.assert_called_with(
+                    flumine=self.mock_flumine,
+                    stream_id=mock_increment(),
+                    market_filter="GANG",
+                    market_data_filter=mock_strategy.market_data_filter,
+                    streaming_timeout=mock_strategy.streaming_timeout,
+                    conflate_ms=mock_strategy.conflate_ms,
+                    output_queue=False,
+                    event_processing=event_processing,
+                    event_group=event_group,
+                    event_id=event_id,
+                    inplay=True,
+                )
 
-    def test_add_historical_stream_old(self):
-        self.mock_flumine.SIMULATED = True
-        mock_strategy = mock.Mock()
-        mock_stream = mock.Mock(event_processing=False, listener_kwargs={})
-        mock_stream.market_filter = "GANG"
-        self.streams._streams = [mock_stream]
-        mock_md = mock.Mock()
+    def test_add_historical_stream_which_already_exists(self):
+        test_args = [
+            # event_processing, event_id, event_group, event_groups
+            # Basic case, no event processing
+            (False, "123", None, {}),
+            # Event groups are ignored since event_processing is False
+            (False, "123", None, {"123": "A"}),
+            # Events groups default to event id if not provided
+            (True, "123", "123", {}),
+            # Event groups are provided
+            (True, "123", "A", {"123": "A"}),
+        ]
+        for subtest_args in test_args:
+            with self.subTest():
+                event_processing, event_id, event_group, event_groups = subtest_args
+                self.mock_flumine.SIMULATED = True
+                mock_strategy = mock.Mock()
+                mock_stream = mock.Mock(
+                    event_processing=event_processing,
+                    event_id=event_id,
+                    event_group=event_group,
+                    listener_kwargs={},
+                )
+                mock_stream.market_filter = "GANG"
+                self.streams._streams = [mock_stream]
 
-        stream = self.streams.add_historical_stream(
-            mock_strategy, "GANG", mock_md, event_processing=False, **{}
-        )
-        self.assertEqual(stream, mock_stream)
-        self.assertEqual(len(self.streams), 1)
+                stream = self.streams.add_historical_stream(
+                    mock_strategy,
+                    "GANG",
+                    mock.Mock(event_id=event_id),
+                    event_processing=event_processing,
+                    event_groups=event_groups,
+                    **{},
+                )
+                self.assertEqual(stream, mock_stream)
+                self.assertEqual(len(self.streams), 1)
 
     @mock.patch("flumine.streams.streams.HistoricalStream")
     @mock.patch("flumine.streams.streams.Streams._increment_stream_id")
@@ -354,14 +396,20 @@ class StreamsTest(unittest.TestCase):
         mock_historical_stream_class.__name__ = "test"
         self.mock_flumine.SIMULATED = True
         mock_strategy = mock.Mock()
-        mock_stream = mock.Mock(event_processing=False, listener_kwargs={})
+        mock_stream = mock.Mock(
+            event_processing=False, event_group=None, listener_kwargs={}
+        )
         mock_stream.market_filter = "GANG"
         self.streams._streams = [mock_stream]
         mock_event_id = "12345"
-        mock_md = mock.Mock(event_id=mock_event_id)
 
         self.streams.add_historical_stream(
-            mock_strategy, "GANG", mock_md, event_processing=False, **{"inplay": True}
+            mock_strategy,
+            "GANG",
+            mock.Mock(event_id=mock_event_id),
+            event_processing=False,
+            event_groups={},
+            **{"inplay": True},
         )
         self.assertEqual(len(self.streams), 2)
         mock_increment.assert_called_with()
@@ -374,6 +422,7 @@ class StreamsTest(unittest.TestCase):
             conflate_ms=mock_strategy.conflate_ms,
             output_queue=False,
             event_processing=False,
+            event_group=None,
             event_id=mock_event_id,
             inplay=True,
         )
@@ -383,13 +432,37 @@ class StreamsTest(unittest.TestCase):
         mock_historical_stream_class.__name__ = "test"
         self.mock_flumine.SIMULATED = True
         mock_strategy = mock.Mock()
-        mock_stream = mock.Mock(event_processing=False)
+        mock_stream = mock.Mock(event_processing=False, event_group=None)
         mock_stream.market_filter = "GANG"
         self.streams._streams = [mock_stream]
-        mock_md = mock.Mock()
-
         stream = self.streams.add_historical_stream(
-            mock_strategy, "GANG", mock_md, event_processing=True
+            mock_strategy,
+            "GANG",
+            mock.Mock(),
+            event_processing=True,
+            event_groups={},
+        )
+        self.assertEqual(stream, mock_historical_stream_class())
+        self.assertEqual(len(self.streams), 2)
+
+    @mock.patch("flumine.streams.streams.get_file_md")
+    @mock.patch("flumine.streams.streams.HistoricalStream")
+    def test_add_historical_stream_event_processing_with_grouping(
+        self, mock_historical_stream_class, mock_get_file_md
+    ):
+        event_id = "123"
+        mock_historical_stream_class.__name__ = "test"
+        self.mock_flumine.SIMULATED = True
+        mock_strategy = mock.Mock()
+        mock_stream = mock.Mock(event_processing=True, event_group=event_id)
+        mock_stream.market_filter = "GANG"
+        self.streams._streams = [mock_stream]
+        stream = self.streams.add_historical_stream(
+            mock_strategy,
+            "GANG",
+            mock.Mock(event_id=event_id),
+            event_processing=True,
+            event_groups={event_id: "A"},
         )
         self.assertEqual(stream, mock_historical_stream_class())
         self.assertEqual(len(self.streams), 2)